+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Version 2.2.0 (2014-xx-xx)
 - added check before deleting non empty categories (see #1)
 - added 'dmsModifyLoaded-Hooks' (see #23)
 - improved handling of file types (see #5)
+- added inheritance of file types (see #5)
 
 IMPORTANT NOTE: activate all access rights after updating vie multi edit mode